# doc/api.yml
openapi: 3.0.3
info:
  title: Birdia API
  description: |
    Detect and qualify geo-referenced objects.
    
    # Security
    
    Get your **API key** by emailing **contact@birdia.fr**.
    All of your requests must embed your API key in the header named: **x-api-key**.
    
    # Usage
    
    Submit your detection request through `POST /detections/{id}`
    where `{id}` is yours to be created.
    
    1. If your zone is described by a **GeoJson**, then just set accordingly the `geoJsonZone` field in the payload.
    2. If your zone is described by a **Shape file**, then upload it in a second call `POST /detections/{id}/shape`.
    3. If your zone is described by a list of addresses in an **Excel file**, then upload it in a second call `POST /detections/{id}/excel`.
    
    # Examples
    
    ### Detection on roof objects providing geojson multiPolygon
    1. Submit detection request through `POST /detections/{id}` where `{id}` is yours to be created
    using the example request body accessible [here](https://raw.githubusercontent.com/b-partners/geo-jobs/refs/heads/preprod/examples/detection/geojson/detection_toiture_payload_example.json).
    Note that `detectableObjectModel` property must be set to `BP_TOITURE` and `geoServerProperties` property to `null`
    2. Track the progress of your detection through `GET /detections/{id}`. 
    Detection is finished when `geoJsonUrl` property containing your detection result is not null.
...
    
...

  version: 'latest'
servers:
...
paths:
...
  '/detectionJobs':
    get:
      tags:
        - MachineDetection
      operationId: getDetectionJobs
      description: List all ZDJ
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        200:
          description: All ZDJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/recomputedStatus':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getRecomputedZDJStatus
      description: Get last recomputed ZDJ status while submitting new recomputing again
      responses:
        200:
          description: The recomputed status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/taskStatistics':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getDetectionTaskStatistics
      description: Get detection task statistics for a given ZDJ
      responses:
        200:
          description: List of detection task statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatistic"
        '500':
          $ref: '#/components/responses/500'
...
  '/detectionJobs/{id}/detectedParcels':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJParcels
      description: List all detected parcels from a ZTJ
      responses:
        200:
          description: All detected parcels for given ZTJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetectedParcel"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/geojsonsUrl':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJGeojsonsUrl
      description: |
        Get a secured URL containing the detection results as geojson.
      responses:
        200:
          description: URL containing detection result as geojson.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoJsonsUrl'
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/process':
    post:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: processZDJ
      description: Process a ZDJ
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DetectableObjectConfiguration"
      responses:
        200:
          description: The processed (-ing) ZDJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'
  '/usage':
    get:
      tags:
        - Usage
      operationId: getDetectionUsage
      summary: Get detection usage
      parameters:
        - in: query
          name: surfaceUnit
          required: false
          schema:
            $ref: '#/components/schemas/DetectionSurfaceUnit'
      responses:
        200:
          description: Statistics of detection usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionUsage'
...
  '/detections':
    get:
      tags:
        - Detection
      operationId: getDetections
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          required: false
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          required: false
      responses:
        200:
          description: List of detections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Detection'
  '/detections/{id}':
    get:
      tags:
        - Detection
      operationId: getProcessedDetection
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      responses:
        200:
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
    post:
      tags:
        - Detection
      operationId: processDetection
      summary: Create or process detection
      description: |
        Create detection and launch it if `geoJsonZone` provided.
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDetection'
            examples:
              detectionForRoof:
                $ref: '#/components/examples/DetectionUsingToitureModel'
      responses:
        200:
          description: Detection created or launched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/step':
    put:
      deprecated: true
      tags:
        - Detection
      operationId: updateDetectionStep
      summary: Update detection step
      description: |
        Update detection step
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectionStep'
      responses:
        200:
          description: Detection with updated step
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/communities/{communityOwnerId}/detections/{id}/step':
    put:
      tags:
        - Detection
      operationId: updateCommunityDetectionStep
      summary: Update detection step for specific community detection
      description: |
        Update detection step for specific community detection
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: communityOwnerId
          schema:
            type: string
          description: Community owner identifier
          required: true
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectionStep'
      responses:
        200:
          description: Detection with updated step
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/addresses':
    post:
      tags:
        - Detection
      summary: Submit detection over a list of textual addresses
      operationId: detectAddresses
      description: |
        Provide list of textual addresses to launch detection
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IdentifierParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Address'
      responses:
        200:
          description: Detection launched with provided textual addresses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/roofer':
    post:
      deprecated: true
      tags:
        - Detection
      operationId: processRooferDetection
      summary: Process detection for roof
      description: |
        This detection is handled synchronously under 30 seconds and used especially for roof under 100 square metre
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDetection'
      responses:
        200:
          description: Detection result for roof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/sync':
    post:
      tags:
        - Detection
      operationId: processDetectionSynchronously
      summary: Process detection synchronously
      description: |
        Intended for small areas such as an address, detection tasks are processed simultaneously, delivering results in less than 60 seconds. 
        Results are requested through the GET /detections/{id} endpoint.
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IdentifierParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDetection'
      responses:
        200:
          description: Detection created or processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  /detections/{id}/roofs/properties:
    put:
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      operationId: computeDetectionRoofsProperties
      responses:
        200:
          description: Detection with computed roofSlope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
        400:
          description: Detection without roofs retrieved exception or Detection without BP_TOITURE model
        404:
          description: Detection not found
  '/detections/{id}/roofer/email':
    post:
      tags:
        - Detection
      operationId: sendMailAboutProspect
      summary: Send information about prospect to roofer
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prospect'
      responses:
        200:
          description: Detection result for roof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/geojson':
    post:
      tags:
        - Detection
      operationId: detectGeoJson
      summary: Submit detection over retrieved geojson
      description: |
        Provide geojson retrieved from shape file or excel containing textual addresses and launch detection
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        description: File containing list of features retrieved
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection launched with updated geojson
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/pdf':
    post:
      tags:
        - Detection
      summary: Submit PDF containing detection result
      operationId: uploadRooferDetectionPdfResult
      description: |
        Provide PDF containing detection result
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        description: Pdf of the detection result
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with pdf file containing result uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/shape':
    post:
      tags:
        - Detection
      operationId: detectShapeFile
      summary: Submit detection over a shape file
      description: |
        Provide shape file to launch detection
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection launched with provided shape file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/excel':
    post:
      tags:
        - Detection
      operationId: detectExcelFile
      summary: Submit detection over an excel file
      description: |
        Provide excel file containing textual addresses and launch detection.
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated excel file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
...
components:
  securitySchemes:
    ReadmeAuth:
      type: apiKey
      name: readme-signature
      in: header
  parameters:
    IdentifierParameter:
      in: path
      name: id
      schema:
        type: string
      description: Detection identifier
      required: true
...
  responses:
...
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
...
    DetectionSurfaceUnit:
      type: string
      enum:
        - SQUARE_DEGREE
...
    ZoneDetectionJob:
      allOf:
        - $ref: "#/components/schemas/CreateZoneDetectionJob"
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/Status"
...
    DetectedParcel:
      properties:
        id:
          type: string
        detectionJobIb:
          type: string
        parcelId:
          type: string
        creationDatetime:
          type: string
          format: date-time
        detectedTiles:
          type: array
          items:
            $ref: "#/components/schemas/DetectedTile"
        status:
          $ref: "#/components/schemas/Status"
    TaskStatistic:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        actualJobStatus:
          $ref: '#/components/schemas/Status'
        taskStatusStatistics:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusStatistic'
        updatedAt:
          type: string
          format: date-time
...
    Status:
      type: object
      properties:
        progression:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - FINISHED
        health:
          type: string
          enum:
            - SUCCEEDED
            - FAILED
            - UNKNOWN
            - RETRYING
        creationDatetime:
          type: string
          format: date-time

    ###########################################################################################################
    ###  GeoJSON objects. Modified from: https://gist.github.com/bubbobne/fe5f2db65acf039be6a9fd92fc9c7233  ###
    ###########################################################################################################
...
    GeoJsonsUrl:
      type: object
      properties:
        url:
          type: string
        status:
          $ref: "#/components/schemas/Status"
...
    Address:
      type: object
      properties:
        address:
          type: string
...